requirements.txt
requests>=2.31.0
pyyaml>=6.0
beautifulsoup4>=4.12.0
trafilatura>=1.6.0